feat: separate API test page
