+++ vite.config.ts
@@ -11,6 +11,14 @@ export default defineConfig(({ mode }) => {
         alias: {
           '@': path.resolve(__dirname, '.'),
         }
+      },
+      build: {
+        rollupOptions: {
+          input: {
+            main: path.resolve(__dirname, 'index.html'),
+            apiTest: path.resolve(__dirname, 'api-test.html'),
+          }
+        }
       }
     };
 });
